+++ test/extension_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: prefer_const_constructors
+
 import 'package:fluent_result/fluent_result.dart';
 import 'package:shouldly/shouldly.dart';
 import 'package:test/test.dart';